Formatting

+++ wacrobot/commands.py
@@ -1,8 +1,8 @@
 import asyncio
 import logging
-from enum import Enum
 
 from bleak import BleakClient
+
 
 class RobotCommand:
     def __init__(self):
@@ -105,37 +105,39 @@ class DisplayDotMatrixCommand(RobotCommand):
         return bytes([0x02] + self.matrix)
 
     async def execute(self, client: BleakClient):
-        logging.debug(f"display text command: {self.matrix} (command: {self.command()})")
+        logging.debug(
+            f"display text command: {self.matrix} (command: {self.command()})"
+        )
         service = client.services.get_service(self._service_uuid)
         char = service.get_characteristic(self._char_uuid)
 
         await client.write_gatt_char(char, self.command(), response=True)
-        logging.debug('sent display text command')
+        logging.debug("sent display text command")
 
 
 class BuzzerCommand(RobotCommand):
     def __init__(self, frequency: int):
         super().__init__()
 
-        self._service_uuid = '1A260001-C2ED-4D11-AD1E-FC06D8A02D37'
-        self._char_uuid = '1A260002-C2ED-4D11-AD1E-FC06D8A02D37'
+        self._service_uuid = "1A260001-C2ED-4D11-AD1E-FC06D8A02D37"
+        self._char_uuid = "1A260002-C2ED-4D11-AD1E-FC06D8A02D37"
 
         self.frequency = frequency
 
     def command(self):
-        return self.frequency.to_bytes(2, 'big')
-    
+        return self.frequency.to_bytes(2, "big")
+
     async def execute(self, client: BleakClient):
         logging.debug(f"buzzer command: {self.frequency}Hz, ({self.command()})")
         service = client.services.get_service(self._service_uuid)
         char = service.get_characteristic(self._char_uuid)
 
         await client.write_gatt_char(char, self.command(), response=True)
-        logging.debug('sent buzzer command')
+        logging.debug("sent buzzer command")
 
 
 class WaitCommand(RobotCommand):
-    def __init__(self, duration: int):
+    def __init__(self, duration: float):
         super().__init__()
         self.duration = duration
